Remove dead code and change structure of file
