+++ api/cmd/api/setup.go
@@ -19,10 +19,10 @@ import (
 	"github.com/gojek/merlin/mlp"
 	"github.com/gojek/merlin/models"
 	"github.com/gojek/merlin/queue"
+	"github.com/gojek/merlin/queue/work"
 	"github.com/gojek/merlin/service"
 	"github.com/gojek/merlin/storage"
 	"github.com/gojek/merlin/vault"
-	"github.com/gojek/merlin/work"
 	"github.com/golang-migrate/migrate/v4"
 	"github.com/golang-migrate/migrate/v4/database/postgres"
 	"github.com/jinzhu/gorm"
