+++ api/service/version_endpoint_service.go
@@ -18,30 +18,16 @@ import (
 	"fmt"
 
 	"github.com/google/uuid"
-	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/gojek/merlin/cluster"
 	"github.com/gojek/merlin/config"
 	"github.com/gojek/merlin/imagebuilder"
 	"github.com/gojek/merlin/log"
-	"github.com/gojek/merlin/mlp"
 	"github.com/gojek/merlin/models"
 	"github.com/gojek/merlin/queue"
+	"github.com/gojek/merlin/queue/work"
 	"github.com/gojek/merlin/storage"
 )
-
-var deploymentCounter = prometheus.NewCounterVec(
-	prometheus.CounterOpts{
-		Name:      "deploy_count",
-		Namespace: "merlin_api",
-		Help:      "Number of deployment",
-	},
-	[]string{"project", "model", "status"},
-)
-
-func init() {
-	prometheus.MustRegister(deploymentCounter)
-}
 
 type EndpointsService interface {
 	ListEndpoints(model *models.Model, version *models.Version) ([]*models.VersionEndpoint, error)
@@ -88,107 +74,6 @@ func NewEndpointService(params EndpointServiceParams) EndpointsService {
 		jobProducer:          params.JobProducer,
 	}
 }
-
-type EndpointJobArgs struct {
-	Endpoint *models.VersionEndpoint
-	Model    *models.Model
-	Version  *models.Version
-	Project  mlp.Project
-}
-
-// ExecuteDeployment will create inferenceservice based on job that already queued
-// func (k *endpointService) ExecuteDeployment(job *queue.Job) error {
-// 	data := job.Arguments[dataArgKey]
-// 	byte, _ := json.Marshal(data)
-// 	var jobArgs models.EndpointJob
-// 	if err := json.Unmarshal(byte, &jobArgs); err != nil {
-// 		return err
-// 	}
-
-// 	endpointArg := jobArgs.Endpoint
-// 	endpoint, err := k.storage.Get(endpointArg.ID)
-// 	if err != nil {
-// 		log.Errorf("could not fetch version endpoint with id %s and error: %v", endpointArg.ID, err)
-// 		return err
-// 	}
-
-// 	version := jobArgs.Version
-// 	project := jobArgs.Project
-// 	model := jobArgs.Model
-
-// 	// Need to reassign destionationURL cause it is ignored when marshalled and unmarshalled
-// 	if endpoint.Logger != nil {
-// 		endpoint.Logger.DestinationURL = k.loggerDestinationURL
-// 	}
-
-// 	model.Project = project
-// 	log.Infof("creating deployment for model %s version %s with endpoint id: %s", model.Name, endpoint.VersionID, endpoint.ID)
-
-// 	// copy endpoint to avoid race condition
-// 	endpoint.Status = models.EndpointFailed
-// 	defer func() {
-// 		deploymentCounter.WithLabelValues(model.Project.Name, model.Name, string(endpoint.Status)).Inc()
-
-// 		// record the deployment result
-// 		if _, err := k.deploymentStorage.Save(&models.Deployment{
-// 			ProjectID:         model.ProjectID,
-// 			VersionModelID:    model.ID,
-// 			VersionID:         endpoint.VersionID,
-// 			VersionEndpointID: endpoint.ID,
-// 			Status:            endpoint.Status,
-// 			Error:             endpoint.Message,
-// 		}); err != nil {
-// 			log.Warnf("unable to insert deployment history", err)
-// 		}
-
-// 		if err := k.storage.Save(endpoint); err != nil {
-// 			log.Errorf("unable to update endpoint status for model: %s, version: %s, reason: %v", model.Name, version.ID, err)
-// 		}
-// 	}()
-
-// 	modelOpt, err := k.generateModelOptions(model, version)
-// 	if err != nil {
-// 		endpoint.Message = err.Error()
-// 		return err
-// 	}
-
-// 	modelService := models.NewService(model, version, modelOpt, endpoint.ResourceRequest, endpoint.EnvVars, k.environment, endpoint.Transformer, endpoint.Logger)
-// 	ctl, ok := k.clusterControllers[endpoint.EnvironmentName]
-// 	if !ok {
-// 		return fmt.Errorf("unable to find cluster controller for environment %s", endpoint.EnvironmentName)
-// 	}
-// 	svc, err := ctl.Deploy(modelService)
-// 	if err != nil {
-// 		log.Errorf("unable to deploy version endpoint for model: %s, version: %s, reason: %v", model.Name, version.ID, err)
-// 		endpoint.Message = err.Error()
-// 		return err
-// 	}
-
-// 	endpoint.URL = svc.URL
-// 	previousStatus := endpointArg.Status
-// 	if previousStatus == models.EndpointServing {
-// 		endpoint.Status = models.EndpointServing
-// 	} else {
-// 		endpoint.Status = models.EndpointRunning
-// 	}
-// 	endpoint.ServiceName = svc.ServiceName
-// 	return nil
-// }
-
-// func (k *endpointService) generateModelOptions(model *models.Model, version *models.Version) (*models.ModelOption, error) {
-// 	modelOpt := &models.ModelOption{}
-// 	switch model.Type {
-// 	case models.ModelTypePyFunc:
-// 		imageRef, err := k.imageBuilder.BuildImage(model.Project, model, version)
-// 		if err != nil {
-// 			return modelOpt, err
-// 		}
-// 		modelOpt.PyFuncImageName = imageRef
-// 	case models.ModelTypePyTorch:
-// 		modelOpt = models.NewPyTorchModelOption(version)
-// 	}
-// 	return modelOpt, nil
-// }
 
 func (k *endpointService) ListEndpoints(model *models.Model, version *models.Version) ([]*models.VersionEndpoint, error) {
 	endpoints, err := k.storage.ListEndpoints(model, version)
@@ -266,7 +151,7 @@ func (k *endpointService) DeployEndpoint(environment *models.Environment, model 
 	if err := k.jobProducer.EnqueueJob(&queue.Job{
 		Name: ModelServiceDeployment,
 		Arguments: queue.Arguments{
-			dataArgKey: models.EndpointJob{
+			dataArgKey: work.EndpointJob{
 				Endpoint: &originalEndpoint,
 				Version:  version,
 				Model:    model,
